Renames gem registry name as omniauth-oidc-strategy
