+++ lib/omniauth-oidc-strategy.rb
